### 2017-05-29 #270 @whedon
[PRE REVIEW]: khmer release v2.1: software for biological sequence analysis
**Submitting author:** @standage (<a href="http://orcid.org/0000-0003-0342-8531">Daniel S. Standage</a>)
**Repository:** <a href="https://github.com/dib-lab/khmer" target ="_blank">https://github.com/dib-lab/khmer</a>
**Version:** v2.1
