### 2017-05-29 #270 @standage
Both @macmanes and @sjackman have been users of khmer (at some point, at least). It appears that neither has reviewed for JOSS before, but were they available they would be able to provide informed reviews for this submission.
