### 2017-05-29 2e078f6
Closes #1706.
