### 2017-05-29 ae60658
Add comment to warning filter for pytest
