+++ pytest.ini
@@ -2,4 +2,7 @@
 addopts = -m "not known_failing and not jenkins and not huge" -v
 norecursedirs = third-party
 testpaths = tests
-filterwarnings = ignore::ResourceWarning
+# ignore all warnings, once khmer stops supporting python2 we can change
+# this to the more specific "ignore::ResourceWarning" to ignore only that
+# one as it is the only one known to be a false positive
+filterwarnings = ignore
