### 2017-05-29 #270 @sjackman
I can review khmer 2.1.
