### 2017-05-30 #272 @arfon
👋  @sjackman - please carry out your review in this issue by updating the checklist above and giving feedback in this issue. The reviewer guidelines are available here: http://joss.theoj.org/about#reviewer_guidelines

Any questions/concerns please let @biorelated know.
