### 2017-05-31 6b3fb5d
Add update_from method to countgraph
