### 2017-06-01 4f5e3f4
Replace ad hoc defines with an enumerated type
