### 2017-06-02 b8326d8
New consume_seqfile_banding method
