### 2017-06-02 07c6146
Add CPython code and tests
