+++ src/khmer/_cpy_hashtable.cc
@@ -110,6 +110,11 @@ PyMethodDef khmer_hashtable_methods[] = {
         "given file"
     },
     {
+        "consume_seqfile_banding",
+        (PyCFunction)hashtable_consume_seqfile_banding, METH_VARARGS,
+        "Consume sequences in k-mer banding mode"
+    },
+    {
         "consume_seqfile_with_reads_parser",
         (PyCFunction)hashtable_consume_seqfile_with_reads_parser, METH_VARARGS,
         "Count all k-mers retrieved with this reads parser object."
@@ -354,6 +359,34 @@ hashtable_consume_seqfile(khmer_KHashtable_Object * me, PyObject * args)
     return Py_BuildValue("IK", total_reads, n_consumed);
 }
 
+PyObject *
+hashtable_consume_seqfile_banding(khmer_KHashtable_Object * me, PyObject * args)
+{
+    Hashtable * hashtable  = me->hashtable;
+
+    const char * filename;
+    unsigned int num_bands;
+    unsigned int band;
+
+    if (!PyArg_ParseTuple(args, "sII", &filename, &num_bands, &band)) {
+        return NULL;
+    }
+
+    // call the C++ function, and trap signals => Python
+    unsigned long long  n_consumed    = 0;
+    unsigned int          total_reads   = 0;
+    try {
+        hashtable->consume_seqfile_banding<FastxReader>(filename, num_bands, band, total_reads, n_consumed);
+    } catch (oxli_file_exception &exc) {
+        PyErr_SetString(PyExc_OSError, exc.what());
+        return NULL;
+    } catch (oxli_value_exception &exc) {
+        PyErr_SetString(PyExc_ValueError, exc.what());
+        return NULL;
+    }
+
+    return Py_BuildValue("IK", total_reads, n_consumed);
+}
 
 PyObject *
 hashtable_consume_seqfile_with_reads_parser(khmer_KHashtable_Object * me,
