+++ src/oxli/kmer_hash.cc
@@ -217,12 +217,15 @@ HashIntoType _hash_murmur_forward(const std::string& kmer, const WordLength k)
 std::pair<uint64_t, uint64_t> compute_band_interval(unsigned int num_bands,
                                                     unsigned int band)
 {
-    if (band < 1 || band > num_bands) {
-        throw InvalidValue("'band' must be in the interval [1, num_bands]");
+    if (band > num_bands) {
+        std::string message = "'band' must be in the interval [0, 'num_bands')";
+        message += ", " + std::to_string(band) + " not in [0, " +
+                   std::to_string(num_bands) + ")";
+        throw InvalidValue(message);
     }
     uint64_t band_size = std::numeric_limits<uint64_t>::max() / num_bands;
-    uint64_t min = band_size * (band - 1);
-    uint64_t max = band_size;
+    uint64_t min = band_size * band;
+    uint64_t max = band_size * (band + 1);
     std::pair<uint64_t, uint64_t> interval (min, max);
     return interval;
 }
