### 2017-06-02 1944633
Cythonize SimpleLabeledAssembler
