+++ tests/test_assembly.py
@@ -59,172 +59,6 @@ def teardown():
     utils.cleanup()
 
 
-class TestLabeledAssembler:
-
-    def test_beginning_to_end_across_tip(self, right_tip_structure):
-        # assemble entire contig, ignoring branch point b/c of labels
-        graph, contig, L, HDN, R, tip = right_tip_structure
-        lh = khmer._GraphLabels(graph)
-        asm = khmer.SimpleLabeledAssembler(lh)
-        hdn = graph.find_high_degree_nodes(contig)
-        # L, HDN, and R will be labeled with 1
-        lh.label_across_high_degree_nodes(contig, hdn, 1)
-
-        path = asm.assemble(contig[:K])
-
-        assert len(path) == 1, "there should only be one path"
-        path = path[0]  # @CTB
-
-        assert len(path) == len(contig)
-        assert utils._equals_rc(path, contig)
-
-    def test_assemble_right_double_fork(self, right_double_fork_structure):
-        # assemble two contigs from a double forked structure
-        graph, contig, L, HDN, R, branch = right_double_fork_structure
-        lh = khmer._GraphLabels(graph)
-        asm = khmer.SimpleLabeledAssembler(lh)
-
-        hdn = graph.find_high_degree_nodes(contig)
-        hdn += graph.find_high_degree_nodes(branch)
-        print(list(hdn))
-        lh.label_across_high_degree_nodes(contig, hdn, 1)
-        lh.label_across_high_degree_nodes(branch, hdn, 2)
-        print(lh.get_tag_labels(list(hdn)[0]))
-
-        paths = asm.assemble(contig[:K])
-        print('Path lengths', [len(x) for x in paths])
-
-        assert len(paths) == 2
-
-        assert any(utils._equals_rc(path, contig) for path in paths)
-        assert any(utils._equals_rc(path, branch) for path in paths)
-
-    def test_assemble_right_triple_fork(self, right_triple_fork_structure):
-        # assemble three contigs from a trip fork
-        (graph, contig, L, HDN, R,
-         top_sequence, bottom_sequence) = right_triple_fork_structure
-        lh = khmer._GraphLabels(graph)
-        asm = khmer.SimpleLabeledAssembler(lh)
-
-        hdn = graph.find_high_degree_nodes(contig)
-        hdn += graph.find_high_degree_nodes(top_sequence)
-        hdn += graph.find_high_degree_nodes(bottom_sequence)
-        print(list(hdn))
-        lh.label_across_high_degree_nodes(contig, hdn, 1)
-        lh.label_across_high_degree_nodes(top_sequence, hdn, 2)
-        lh.label_across_high_degree_nodes(bottom_sequence, hdn, 3)
-        print(lh.get_tag_labels(list(hdn)[0]))
-
-        paths = asm.assemble(contig[:K])
-        print([len(x) for x in paths])
-
-        assert len(paths) == 3
-
-        assert any(utils._equals_rc(path, contig) for path in paths)
-        assert any(utils._equals_rc(path, top_sequence) for path in paths)
-        assert any(utils._equals_rc(path, bottom_sequence) for path in paths)
-
-    def test_assemble_left_double_fork(self, left_double_fork_structure):
-        # assemble entire contig + branch points b/c of labels; start from end
-        graph, contig, L, HDN, R, branch = left_double_fork_structure
-        lh = khmer._GraphLabels(graph)
-        asm = khmer.SimpleLabeledAssembler(lh)
-
-        # first try without the labels
-        paths = asm.assemble(contig[-K:])
-
-        assert len(paths) == 1
[…]
