+++ khmer/_oxli/assembly.pyx
@@ -64,3 +64,35 @@ cdef class LinearAssembler:
             return self._assemble_right(Kmer(str(seed)))
 
 
+cdef class SimpleLabeledAssembler:
+
+    def __cinit__(self, labels, stop_filter=None):
+        self._label_ptr = get_labelhash_ptr(labels)
+        self.labels = labels
+        self.set_stop_filter(stop_filter=stop_filter)
+        
+        if type(self) is SimpleLabeledAssembler:
+            self._this.reset(new CpSimpleLabeledAssembler(self._label_ptr))
+
+    def set_stop_filter(self, stop_filter=None):
+        self.stop_filter = stop_filter
+        if stop_filter is not None:
+            self._stop_filter_ptr = get_hashgraph_ptr(stop_filter)
+            if self._stop_filter_ptr == NULL:
+                raise ValueError('Must take an object with Hashgraph *')
+        else:
+            self._stop_filter_ptr = NULL
+
+    cdef vector[string] _assemble(self, Kmer kmer):
+        if self.stop_filter is None:
+            return deref(self._this).assemble(deref(kmer._this))
+        else:
+            return deref(self._this).assemble(deref(kmer._this), self._stop_filter_ptr)
+
+    def assemble(self, seed):
+        if isinstance(seed, Kmer):
+            return self._assemble(seed)
+        else:
+            return self._assemble(Kmer(str(seed)))
+
+
