### 2017-06-02 67202ec
Cythonize JunctionCountAssembler and remove old files
