+++ src/khmer/_cpy_assembly.cc
@@ -1,185 +1,0 @@
-#include "khmer/_cpy_assembly.hh"
-#include "khmer/_cpy_nodegraph.hh"
-#include "khmer/_cpy_countgraph.hh"
-#include "khmer/_cpy_graphlabels.hh"
-
-using namespace oxli;
-
-namespace khmer {
-
-
-/********************************
- * JunctionCountAssembler
- ********************************/
-
-
-PyTypeObject khmer_KJunctionCountAssembler_Type = {
-    PyVarObject_HEAD_INIT(NULL, 0)  /* init & ob_size */
-    "_khmer.JunctionCountAssembler",            /* tp_name */
-    sizeof(khmer_KJunctionCountAssembler_Object), /* tp_basicsize */
-    0,                       /* tp_itemsize */
-    (destructor)khmer_junctioncountassembler_dealloc, /* tp_dealloc */
-    0,                       /* tp_print */
-    0,                       /* tp_getattr */
-    0,                       /* tp_setattr */
-    0,                       /* tp_compare */
-    0,                       /* tp_repr */
-    0,                       /* tp_as_number */
-    0,                       /* tp_as_sequence */
-    0,                       /* tp_as_mapping */
-    0,                       /* tp_hash */
-    0,                       /* tp_call */
-    0,                       /* tp_str */
-    0,                       /* tp_getattro */
-    0,                       /* tp_setattro */
-    0,                       /* tp_as_buffer */
-    Py_TPFLAGS_DEFAULT | Py_TPFLAGS_BASETYPE,   /* tp_flags */
-    0,                       /* tp_doc */
-    0,                       /* tp_traverse */
-    0,                       /* tp_clear */
-    0,                       /* tp_richcompare */
-    0,                       /* tp_weaklistoffset */
-    0,                       /* tp_iter */
-    0,                       /* tp_iternext */
-    khmer_junctioncountassembler_methods, /* tp_methods */
-    0,                       /* tp_members */
-    0,                       /* tp_getset */
-    0,                       /* tp_base */
-    0,                       /* tp_dict */
-    0,                       /* tp_descr_get */
-    0,                       /* tp_descr_set */
-    0,                       /* tp_dictoffset */
-    0,                       /* tp_init */
-    0,                       /* tp_alloc */
-    khmer_junctioncountassembler_new,      /* tp_new */
-};
-
-PyMethodDef khmer_junctioncountassembler_methods[] = {
-    {
-        "assemble",
-        (PyCFunction)junctioncountassembler_assemble, METH_VARARGS | METH_KEYWORDS,
-        "Assemble paths, using recorded junctions to jump branches."
-    },
-    {
-        "consume",
-        (PyCFunction)junctioncountassembler_consume, METH_VARARGS,
-        "Consume a string and count its branch junctions."
-    },
-    {NULL, NULL, 0, NULL}           /* sentinel */
-};
-
-
-void khmer_junctioncountassembler_dealloc(
-    khmer_KJunctionCountAssembler_Object * obj)
-{
-    delete obj->assembler;
-    obj->assembler = NULL;
-
-    Py_TYPE(obj)->tp_free((PyObject*)obj);
-}
-
-PyObject * khmer_junctioncountassembler_new(PyTypeObject *type,
-        PyObject *args,
-        PyObject *kwds)
-{
-    khmer_KJunctionCountAssembler_Object *self;
-    self = (khmer_KJunctionCountAssembler_Object*)type->tp_alloc(type, 0);
-
-    if (self != NULL) {
-        PyObject * hashgraph_o;
[…]
