+++ khmer/__init__.py
@@ -51,7 +51,6 @@ from khmer._khmer import Nodetable as _Nodetable
 from khmer._khmer import HLLCounter as _HLLCounter
 from khmer._khmer import ReadAligner as _ReadAligner
 
-from khmer._khmer import JunctionCountAssembler
 from khmer._khmer import HashSet
 from khmer._khmer import Read
 from khmer._khmer import forward_hash
@@ -515,4 +514,5 @@ class ReadAligner(_ReadAligner):
         """
         _ReadAligner.__init__(self)
 
-from khmer._oxli.assembly import LinearAssembler, SimpleLabeledAssembler
+from khmer._oxli.assembly import (LinearAssembler, SimpleLabeledAssembler,
+                                  JunctionCountAssembler)
