### 2017-06-02 03cebec
CHANGELOG and num_bands != 2^N in tests
