### 2017-06-07 #272 @sjackman
@standage Could the DOIs be made into links using http://dx.doi.org ? https://github.com/dib-lab/khmer/blob/master/papers/paper-v2.1.md#references
