### 2017-06-07 dib-lab/khmer#1722 @sjackman
load-into-counting.py: ValueError: bigcount is not supported for this storage
`load-into-counting.py` with `--small-count` yields `ValueError: bigcount is not supported for this storage.`. Adding `--no-bigcount` doesn't seem to help.

```
❯❯❯ load-into-counting.py --small-count -x 1e7 khmer2 foo.fq.gz
PARAMETERS:
 - kmer size =     32 		(-k)
 - n tables =      4 		(-N)
 - max tablesize = 1e+07 	(-x)
Estimated memory usage is 0.0 Gb (4e+07 bytes = 4 bytes x 1e+07 entries / 1 entries per byte)
--------
Saving k-mer countgraph to khmer2
Loading kmers from sequences in ['foo.fq.gz']
making countgraph
Traceback (most recent call last):
  File "/Users/sjackman/.homebrew/bin/load-into-counting.py", line 221, in <module>
    main()
  File "/Users/sjackman/.homebrew/bin/load-into-counting.py", line 139, in main
    countgraph.set_use_bigcount(args.bigcount)
ValueError: bigcount is not supported for this storage.
❯❯❯ load-into-counting.py --version
khmer 2.1.1
```
