### 2017-06-07 #272 @sjackman
I tried to verify the memory usage of `--small-count` and ran into an issue, reported at https://github.com/dib-lab/khmer/issues/1722
