### 2017-06-07 dib-lab/khmer#1722 @betatim
I will take a look at this.
