### 2017-06-07 747cabf
Add tests for small-count vs bigcount CLI conflict
The two arguments conflict with each other as you can not set
"bigcount" on a graph or table that uses smallcount. This adds
tests and modifies the code to check if smallcount is set or
not.
