### 2017-06-07 dib-lab/khmer#1722 @betatim
Thanks for catching this!
