### 2017-06-07 #272 @standage
> Could the DOIs be made into links using http://dx.doi.org?

Sure, I'll take care of that.
