### 2017-06-07 dib-lab/khmer#1725 @sjackman
examples/stamps/do.sh: error: unrecognized arguments: --savetable stamps-dn.ct
It looks as though `--savetable` should be either `-s` or `--savegraph`.
See https://github.com/dib-lab/khmer/blob/stable/examples/stamps/do.sh#L43
and https://github.com/dib-lab/khmer/blob/stable/examples/stamps/do.sh#L60
