### 2017-06-07 dib-lab/khmer#1722 @sjackman
No worries! Is there a workaround for version 2.1.1? `load-into-counting.py --small-count --no-bigcount -x 1e7 khmer2 foo.fq.gz` didn't work for me.
