### 2017-06-07 dib-lab/khmer#1722 @sjackman
Or can you suggest a different command that I can use to test the memory usage of `--small-count` vs the default?
