### 2017-06-07 575dd8f
Fix counting options bug a la #1723.
