### 2017-06-07 dib-lab/khmer#1722 @standage
I'm porting this to the v2.1 maintenance branch now. Will cut a bugfix release once the JOSS review is complete.
