### 2017-06-07 #272 @sjackman
I tried to run `examples/stamps/do.sh` and encountered the error `unrecognized arguments: --savetable`, reported at https://github.com/dib-lab/khmer/issues/1725
