### 2017-06-07 dib-lab/khmer#1722 @sjackman
For the item `Performance: Have any performance claims of the software been confirmed?` I'm trying to confirm the reduced memory usage of `--small-count` vs the default. Can you suggest a command?
