### 2017-06-07 8361ac7
Fix missing graphtype declaration for config printout
