### 2017-06-07 dib-lab/khmer#1722 @standage
Once #1726 is merged, you should be able to run the following commands on branch `maint/2.1` (from which release version 2.1.2 will be cut when the JOSS review is complete).

## Same graph size, same FPR, less memory

```
scripts/load-into-counting.py -x 2e6 default.count tests/test-data/test-reads.fq.gz
scripts/load-into-counting.py -x 2e6 small.count --small-count tests/test-data/test-reads.fq.gz
```

The reduced memory usage should be reflected in the configuration dump printed to the screen, as well as the count graph file sizes (`default.cg` vs `small.cg`).

## Same amount of memory, larger graph size, smaller FPR

```
load-into-counting.py -M 4M default.cg tests/test-data/test-reads.fq.gz
load-into-counting.py -M 4M small.cg --small-count tests/test-data/test-reads.fq.gz
```

The output file sizes should be the same, but the reported table size for the `--small-count` should be larger, and the reported FPR smaller.
