### 2017-06-07 dib-lab/khmer#1722 @sjackman
Thanks. I've subscribed to PR #1726.
