### 2017-06-07 #272 @standage
Links fixed in https://github.com/dib-lab/khmer/pull/1727.
