### 2017-06-07 dib-lab/khmer#1725 @standage
Curious. I was unable to reproduce this, and none of tag `v2.1.1`, branch `maint/2.1`, or branch `master` have `--savetable` in `examples/stamps/do.sh`. Are you sure there's not a conflict with an older version of khmer installed system-wide?
