### 2017-06-07 dib-lab/khmer#1725 @standage
Ooops, looks like we have a `stable` tag that is, uh, not stable. :-/

This explains some weird behavior we've been seeing with our doc builds recently.
