### 2017-06-08 eb4a9d3
Add missed namespace renames and return type change
