### 2017-06-08 c626837
Experimenting with adding a cython interface
