+++ setup.py
@@ -307,8 +307,8 @@ class KhmerBuildExt(_build_ext):  # pylint: disable=R0904
 
         cqfcmd = ['bash', '-c', 'cd third-party/cqf && make']
         spawn(cmd=cqfcmd, dry_run=self.dry_run)
-        self.extensions[0].extra_objects.append(
-            path_join("third-party", "cqf", "gqf.o"))
+        for ext in self.extensions:
+            ext.extra_objects.append(path_join("third-party", "cqf", "gqf.o"))
 
         if "z" not in self.libraries:
             zcmd = ['bash', '-c', 'cd ' + ZLIBDIR + ' && ( test Makefile -nt'
