### 2017-06-09 dib-lab/khmer#1722 @sjackman
Your suggested commands worked on the `maint/2.1` branch. Thanks.
