### 2017-06-09 1657f70
Add pytest-runner as dev package dependency
