### 2017-06-09 dib-lab/khmer#1722 @standage
Excellent.
