### 2017-06-09 ed41ad3
Kind of working, still need to raise appropriate exceptions
