### 2017-06-11 52eac78
Add exception handling
