+++ setup.py
@@ -154,15 +154,16 @@ BZIP2DIR = 'third-party/bzip2'
 BUILD_DEPENDS = glob.glob(path_join("include", "khmer", "_cpy_*.hh"))
 BUILD_DEPENDS.extend(path_join("include", "oxli", bn + ".hh") for bn in [
     "khmer", "kmer_hash", "hashtable", "labelhash", "hashgraph",
-    "hllcounter", "khmer_exception", "read_aligner", "subset", "read_parsers",
-    "kmer_filters", "traversal", "assembler", "alphabets", "storage"])
+    "hllcounter", "oxli_exception", "read_aligner", "subset", "read_parsers",
+    "kmer_filters", "traversal", "assembler", "alphabets", "storage",
+    "oxli_exception_convert"])
 
 SOURCES = glob.glob(path_join("src", "khmer", "_cpy_*.cc"))
 SOURCES.extend(path_join("src", "oxli", bn + ".cc") for bn in [
     "read_parsers", "kmer_hash", "hashtable", "hashgraph",
     "labelhash", "subset", "read_aligner",
     "hllcounter", "traversal", "kmer_filters", "assembler", "alphabets",
-    "storage"])
+    "storage", "oxli_exception_convert"])
 
 SOURCES.extend(path_join("third-party", "smhasher", bn + ".cc") for bn in [
     "MurmurHash3"])
