### 2017-06-11 80d0b4f
Fix error rate issues
