### 2017-06-11 f6977cf
Fix for py27
