### 2017-06-12 6fbfb81
Add docstrings to basic cython interface
