+++ khmer/_oxli/graphs.pyx
@@ -3,6 +3,7 @@ from cython.operator cimport dereference as deref
 
 from libcpp.memory cimport unique_ptr
 
+from .utils cimport _bstring
 from .._khmer import Countgraph, Nodegraph, GraphLabels
 
 
@@ -22,21 +23,52 @@ cdef CpLabelHash * get_labelhash_ptr(object labels):
     return deref(ptr).labelhash
 
 
-cdef class QFCounttable_t:
+cdef class QFCounttable_:
     cdef unique_ptr[CpQFCounttable] c_table
     def __cinit__(self, int k, int size):
         self.c_table.reset(new CpQFCounttable(k, size))
 
     def add(self, kmer):
-        if not isinstance(kmer, unicode):
-            raise ValueError("requires text input, got %s" % type(kmer))
-        bytes = kmer.encode("ASCII")
-        cdef const char* data = bytes
-        return deref(self.c_table).add(data)
+        """Increment the count of this k-mer.
 
-    def get_count(self, kmer):
-        if not isinstance(kmer, unicode):
-            raise ValueError("requires text input, got %s" % type(kmer))
-        bytes = kmer.encode("ASCII")
-        cdef const char* data = bytes
-        return deref(self.c_table).get_count(data)
+        Synonym for 'count'.
+        """
+        return self.count(kmer)
+
+    def count(self, kmer):
+        """Increment the count of this k-mer
+
+        `kmer` can be either a string or an integer representing the hashed
+        value of the kmer.
+        """
+        if isinstance(kmer, (unicode, str)):
+            data = _bstring(kmer)
+            return deref(self.c_table).add(data)
+        # assume kmer is an integer representing the hash value
+        else:
+            return deref(self.c_table).add(kmer)
+
+    def hash(self, kmer):
+        """"Returns the hash of this k-mer.
+
+        For Nodetables and Counttables, this function will fail if the
+        supplied k-mer contains non-ACGT characters.
+        """
+        data = _bstring(kmer)
+        return deref(self.c_table).hash_dna(data)
+
+    def get(self, kmer):
+        """"Retrieve the count for the given k-mer.
+
+        `kmer` can be either a string or an integer representing the hashed
+        value of the kmer.
+
+        For Nodetables and Counttables, this function will fail if the
+        supplied k-mer contains non-ACGT characters.
+        """
+        if isinstance(kmer, (unicode, str)):
+            data = _bstring(kmer)
+            return deref(self.c_table).get_count(data)
+        # assume kmer is an integer representing the hash value
+        else:
+            return deref(self.c_table).get_count(kmer)
