### 2017-06-12 e5f5898
Handle overloading by hand
