+++ khmer/_oxli/graphs.pyx
@@ -3,9 +3,11 @@ from cython.operator cimport dereference as deref
 from libc.stdint cimport uint64_t
 
 from libcpp.memory cimport unique_ptr
+from libcpp.vector cimport vector
 
 from .utils cimport _bstring
 from graphs cimport CpQFCounttable
+from parsing cimport CpFastxReader, FastxParserPtr
 from .._khmer import Countgraph, Nodegraph, GraphLabels
 
 
@@ -51,16 +53,19 @@ cdef class QFCounttable_:
             return deref(self.c_table).add(<uint64_t>kmer)
 
     def hash(self, kmer):
-        """"Returns the hash of this k-mer.
-
-        For Nodetables and Counttables, this function will fail if the
-        supplied k-mer contains non-ACGT characters.
-        """
+        """Compute the hash of this k-mer."""
+        if len(kmer) != self.ksize():
+            raise ValueError("Expected k-mer length {}"
+                             " but got {}.".format(self.ksize(), len(kmer)))
         data = _bstring(kmer)
         return deref(self.c_table).hash_dna(data)
 
+    def reverse_hash(self, kmer_hash):
+        """Turn a k-mer hash back into a DNA k-mer, if possible."""
+        return deref(self.c_table).unhash_dna(kmer_hash)
+
     def get(self, kmer):
-        """"Retrieve the count for the given k-mer.
+        """Retrieve the count for the given k-mer.
 
         `kmer` can be either a string or an integer representing the hashed
         value of the kmer.
@@ -74,3 +79,97 @@ cdef class QFCounttable_:
         # assume kmer is an integer representing the hash value
         else:
             return deref(self.c_table).get_count(<uint64_t>kmer)
+
+    def ksize(self):
+        """k-mer size"""
+        return deref(self.c_table).ksize()
+
+    def hashsizes(self):
+        """Size of hash tables used."""
+        return deref(self.c_table).get_tablesizes()
+
+    cdef _valid_sequence(self, sequence):
+        """Validate sequence argument and convert it to bytes"""
+        if len(sequence) < self.ksize():
+            raise ValueError("sequence length ({}) must >= the hashtable "
+                             "k-mer size ({})".format(len(sequence),
+                                                      self.ksize()))
+        return _bstring(sequence)
+
+    def get_kmers(self, sequence):
+        """Generate an ordered list of all k-mers in sequence."""
+        data = self._valid_sequence(sequence)
+        cdef vector[string] kmers
+        deref(self.c_table).get_kmers(data, kmers)
+        return kmers
+
+    def consume(self, sequence):
+        """Increment the counts of all of the k-mers in the sequence."""
+        data = self._valid_sequence(sequence)
+        return deref(self.c_table).consume_string(data)
+
+    def get_kmer_counts(self, sequence):
+        """Retrieve an ordered list of the counts of all k-mers in sequence."""
+        data = self._valid_sequence(sequence)
+        cdef vector[BoundedCounterType] counts
+        deref(self.c_table).get_kmer_counts(data, counts)
+        return counts
+
+    def get_min_count(self, sequence):
+        """Get the smallest count of all the k-mers in the string."""
+        data = self._valid_sequence(sequence)
+        return deref(self.c_table).get_min_count(data)
+
+    def get_max_count(self, sequence):
+        """Get the larget count of all the k-mers in the string."""
+        data = self._valid_sequence(sequence)
+        return deref(self.c_table).get_max_count(data)
+
+    def get_median_count(self, sequence):
+        """median, average, and stddev of the k-mer counts in sequence."""
+        data = self._valid_sequence(sequence)
+        cdef BoundedCounterType med = 0
+        cdef float average = 0
+        cdef float stddev = 0
+
[…]
