### 2017-06-13 60e7210
Use smaller input file for banding tests
