### 2017-06-14 fd2cebb
Experiment with ReadParser support
