### 2017-06-14 dib-lab/khmer#1725 @standage
Stable tag was removed.
