### 2017-06-14 fe4db4f
Better solution
