### 2017-06-15 8ba9094
Switch to cython reads parser
