### 2017-06-15 3ac01e6
Add abundance_distribution method
