### 2017-06-15 #272 @sjackman
> support for k > 32 using the non-reversible hash function MurmurHash3

Is there an upper limit on the value of *k*?

Consider italicizing each occurrence of *k*, e.g. *k*-mer and *k* > 32.
