### 2017-06-15 #272 @standage
> Is there an upper limit on the value of *k*?

Nope, MurmurHash3 will handle arbitrarily large inputs. Of course, the larger *k* gets the more likely collisions in *k*-mer hash space are. But this is offset by two factors: 1) *k*-mers observed in genomes are sparse with respect to the entire *k*-mer space, especially as *k* gets larger; 2) collisions in *k*-mer hash space are orders of magnitude less frequent than collisions in the probabilistic data structures we use.

> Consider italicizing each occurrence of *k*, e.g. *k*-mer and *k* > 32.

Done. Thanks for catching that!
