### 2017-06-15 #272 @sjackman
```
---
title: 'Review of khmer release v2.1: software for biological sequence analysis'
author: Shaun Jackman
date: 2017-06-15
---
```

The authors describe version 2.1 of their software khmer for various analyses of sequencing data. khmer implements primarily two data structures, the Bloom filter to store *k*-mer presence/absence and Count-Min Sketch to store *k*-mer counts. This release adds new features including variable-coverage trimming, supporting *k* > 32, reduced memory utilization for the Count-Min Sketch by using 4 bit counters rather than 8, and assembling and compacting a de Bruijn graph.

The software installed easily using `pip install khmer`. I successfully ran the example `stamps` analysis of a synthetic metagenomic data set. I verified the 4-bit counter reduces the memory usage of khmer by half. I successfully ran the de Bruijn graph compaction script `extract-compact-dbg.py`.

The 4-bit counter will be useful in reducing the memory requirements of assembling large genomes and metagenomes. Values of *k* larger than 32 will be useful in assembling 150-bp reads, where values of *k* around 100 are common. The de Bruijn graph assembly and compaction scripts are a useful contribution to the field of de Bruin graph assembly using compact data structures.

Although likely out of scope for this paper, I would like to see
1. the assembly modules optionally output the assembled graph in the Graphical Fragment Assembly (GFA) file format, or provide a converter from the default GML output format to GFA for interoperability with other assembly tools and visualization with Bandage.
2. a description of the de Bruin graph assembly and compaction algorithms.
3. a comparison of the assembly feature of khmer to other compact data structure assemblers like Minia, BCALM, and ABySS 2.0.
