### 2017-06-15 #272 @sjackman
@biorelated Hi, George. I have completed my review of khmer 2.1.
