### 2017-06-16 bcdb1a8
Move __new__ code to __cinit__
