### 2017-06-16 #272 @biorelated
Thank you very much @sjackman  for taking time to review. You have raised a couple of points that you would like to see although they might be out of scope of  this manuscript. @standage could you confirm whether these are items that you would like to add to the current software package (and manuscript) or it is something that you would like to implement in a future release.
