### 2017-06-16 401e990
Switch to using ReadParser everywhere
