### 2017-06-16 e8fe765
Remove log2() for python2 compatibility
