### 2017-06-16 #272 @standage
Thanks for your comments and your generosity with time, @sjackman and @biorelated!

Support for GFA output is something we've been discussing recently (see https://github.com/dib-lab/khmer/issues/1467), but this didn't get prioritized for the 2.1 release.
I'm hoping it's something we can include by version 3.0.

The graph compaction code is convenient but also intentionally unsophisticated; it’s not interesting enough for a research paper.
We’ll document it as we expand the Python interface, probably for 3.0

A description of khmer's assembly algorithm and comparisons to other tools, while potentially insightful from a scientific perspective, is indeed outside the scope of this software paper.
At least one member of our research group is actively developing novel assembly methods and applications, and publication of these developments should provide an opportunity to describe khmer's implicit de Bruijn graph implementation and graph traversal algorithms in depth.

Best wishes,
Daniel Standage
