### 2017-06-17 #272 @biorelated
Thank you Daniel (@standage)  for the quick response and thank you again Shaun (@sjackman) for your review. @arfon please could we get this ready for publication?
Thank you
George
