### 2017-06-17 #272 @arfon
>  @arfon please could we get this ready for publication?

👍

@standage - Could you move the references you currently have in the `paper-v2.1.md` file into a `paper.bib` file and cite them directly please? (You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html))
