### 2017-06-19 a2200ed
Add load/save to QFCounttable
