+++ tests/test_tabletype.py
@@ -23,7 +23,7 @@ from . import khmer_tst_utils as utils
 import khmer
 from khmer import _Countgraph, _Counttable, _SmallCountgraph, _SmallCounttable
 from khmer import _Nodegraph, _Nodetable, QFCounttable
-from khmer import ReadParser, FastxParser
+from khmer import ReadParser
 import screed
 
 
