### 2017-06-19 d942533
Add CQF to liboxli (fixes make cpp-demo)
