### 2017-06-20 5246cb1
Use separate bibliography file for v2.1 JOSS pub
