+++ papers/paper-v2.1.md
@@ -93,29 +93,15 @@ date: 2017-05-25
 # Summary
 
 The khmer software is a set of command-line tools built around a Python library designed for analysis of large DNA sequence collections.
-Functionality in khmer has primarily been motivated by scaling issues with (meta)genome and (meta)transcriptome assembly (Crusoe *et al.*, 2015).
-khmer provides convenient access to several *k*-mer based operations on DNA sequence collections, such as abundance filtering, error trimming, assembly graph partitioning, and most notably, abundance normalization of reads (Brown *et al.*, 2012) and streaming error trimming of reads (Zhang, Awad, and Brown, 2015).
-All of these operations utilize khmer's implementation of two primary data structures, the Bloom filter and the Count-Min Sketch, for efficient probabalistic storage of *k*-mer presence or *k*-mer abundance, respectively (Pell *et al.*, 2012; Zhang *et al.*, 2014).
+Functionality in khmer has primarily been motivated by scaling issues with (meta)genome and (meta)transcriptome assembly [@crusoe2015].
+khmer provides convenient access to several *k*-mer based operations on DNA sequence collections, such as abundance filtering, error trimming, assembly graph partitioning, and most notably, abundance normalization of reads [@brown2012] and streaming error trimming of reads [@zhang2015].
+All of these operations utilize khmer's implementation of two primary data structures, the Bloom filter and the Count-Min Sketch, for efficient probabalistic storage of *k*-mer presence or *k*-mer abundance, respectively [@pell2012; @zhang2014].
 
 Release version 2.1 of the khmer software includes several new features that extend its utility to a wider set of sequence processing and analysis problems.
 These include the following:
 support for variable-coverage trimming of sequence reads;
-support for *k* > 32 using the non-reversible hash function MurmurHash3 (Appleby, 2010);
+support for *k* > 32 using the non-reversible hash function MurmurHash3 (https://github.com/aappleby/smhasher);
 a new optional Count-Min Sketch implementation providing increased storage efficiency;
 support for assembly directly from a *k*-mer graph;
 a script for computing a compact de Bruijn graph from a *k*-mer graph;
 and several examples of khmer's Python and C++ APIs for those interested in using and extending the library.
-
-# References
-
-**Appleby A** (2010) SMHasher. *GitHub*, https://github.com/aappleby/smhasher, accessed May 22, 2017.
-
-**Brown CT, Howe A, Zhang Q, Pyrkosz AB, Brom TH** (2012) A Reference-Free Algorithm for Computational Normalization of Shotgun Sequencing Data. *arXiv*, arXiv:1203.4802, https://arxiv.org/abs/1203.4802.
-
-**Crusoe MR, Alameldin HF, Awad S, Bucher E, *et al.*** (2015) The khmer software package: enabling efficient nucleotide sequence analysis. *F1000 Research*, [doi:10.12688/f1000research.6924.1](http://dx.doi.org/10.12688/f1000research.6924.1).
-
-**Pell J, Hintze A, Canino-Koning R, Howe A, Tiedje JM, Brown CT** (2012). Scaling metagenome sequence assembly with probabilistic de Bruijn graphs. *Proc Natl Acad Sci U S A*, 109(33):13272-7, [doi:10.1073/pnas.1121464109](http://dx.doi.org/10.1073/pnas.1121464109).
-
[…]
