### 2017-06-20 #272 @standage
> Could you move the references you currently have in the paper-v2.1.md file into a paper.bib file and cite them directly please?

Done. See latest commit in `maint/2.1` branch.
