### 2017-06-23 a223939
test Dockerfile and add Makefile target
