+++ docker/Dockerfile
@@ -1,6 +1,5 @@
 FROM debian:stable
 MAINTAINER khmer-project@idyll.org
-ARG branch=v2.0
 
 ENV PACKAGES python-dev zlib1g-dev libbz2-dev gcc git python-setuptools g++ make ca-certificates python-pip python-wheel 
 
@@ -10,7 +9,10 @@ RUN apt-get update && \
 
 WORKDIR /home
 
-RUN git clone --branch $branch https://github.com/dib-lab/khmer.git && \
+ARG branch=v2.0
+
+RUN git clone https://github.com/dib-lab/khmer.git && \
     cd khmer && \
+    git checkout $branch \
     make install-dep && \
     python setup.py build_ext --include-dirs third-party/seqan/core/include:third-party/smhasher:lib --libraries z,bz2 install
