### 2017-06-23 ff8cf31
just build Docker container under Linux
