+++ .travis.yml
@@ -72,7 +72,7 @@ script:
   - make py-demos
   - make cpp-demos
   - cd examples/stamps && ./do.sh
-  - make docker-container
+  - if [[ "$TRAVIS_OS_NAME" == "linux" ]]; then cd .. && make docker-container
 
 # generate all the diagnostic reports
 after_success:
