### 2017-06-23 a6b8115
fix py3 compat for setup.py
