### 2017-06-23 7b771ef
Add methods to query n_occupied slots in CQF storage
