### 2017-06-23 f5781ac
Add CHANGELOG message
