### 2017-06-23 6b7ab9b
get real commit hash from Travis
