### 2017-06-25 #272 @arfon
@standage - there's still some problems here sorry.

1. Please link to the `paper.bib` file in the YAML header: `bibliography: paper.bib` (see the example paper here: https://raw.githubusercontent.com/arfon/fidgit/master/paper/paper.md)
2. Please add a `# References` section to the `paper-v2.1.md` file (this is where the pandoc compilation adds your BibTeX references automatically)
3. Please make sure you're citing references in your paper that you actually have in your BibTeX. Currently a number are missing:
```
pandoc-citeproc: reference crusoe2015 not found
pandoc-citeproc: reference zhang2015 not found
pandoc-citeproc: reference pell2012 not found
pandoc-citeproc: reference zhang2014 not found
```
