### 2017-06-25 0bf0a70
Explicitly specify utf8 encoding for authors.csv
