### 2017-06-25 42f0930
Different encoding strings for python2 and python3 for authors.csv
