### 2017-06-26 913fdb2
Fix bibliography
