### 2017-06-26 #272 @standage
Updated. Thanks @arfon.

The text already cited those references as is. Perhaps the message appeared because I had not properly included the `.bib` file and/or included a header for the **References** section?

Hopefully this addresses it. Let me know if anything is still amiss.
