### 2017-06-26 2df0110
Update paper.bib
