+++ papers/paper.bib
@@ -5,7 +5,7 @@ Year = {2012},
 Eprint = {arXiv:1203.4802},
 }
 
-@Article{Crusoe2015,
+@Article{crusoe2015,
 AUTHOR = { Crusoe, MR and Alameldin, HF and Awad, S and Boucher, E and Caldwell, A and Cartwright, R and Charbonneau, A and Constantinides, B and Edvenson, G and Fay, S and Fenton, J and Fenzl, T and Fish, J and Garcia-Gutierrez, L and Garland, P and Gluck, J and Gonzalez, I and Guermond, S and Guo, J and Gupta, A and Herr, JR and Howe, A and Hyer, A and H�rpfer, A and Irber, L and Kidd, R and Lin, D and Lippi, J and Mansour, T and McA'Nulty, P and McDonald, E and Mizzi, J and Murray, KD and Nahum, JR and Nanlohy, K and Nederbragt, AJ and Ortiz-Zuazaga, H and Ory, J and Pell, J and Pepe-Ranney, C and Russ, ZN and Schwarz, E and Scott, C and Seaman, J and Sievert, S and Simpson, J and Skennerton, CT and Spencer, J and Srinivasan, R and Standage, D and Stapleton, JA and Steinman, SR and Stein, J and Taylor, B and Trimble, W and Wiencko, HL and Wright, M and Wyss, B and Zhang, Q and zyme, e and Brown, CT},
 TITLE = {The khmer software package: enabling efficient nucleotide sequence analysis},
 JOURNAL = {F1000Research},
@@ -15,7 +15,7 @@ NUMBER = {900},
 DOI = {10.12688/f1000research.6924.1}
 }
 
-@article{Pell2011,
+@article{pell2012,
 author = {Pell, Jason and Hintze, Arend and Canino-Koning, Rosangela and Howe, Adina and Tiedje, James M. and Brown, C. Titus},
 title = {Scaling metagenome sequence assembly with probabilistic de Bruijn graphs},
 volume = {109},
@@ -28,7 +28,7 @@ eprint = {http://www.pnas.org/content/109/33/13272.full.pdf},
 journal = {Proceedings of the National Academy of Sciences}
 }
 
-@article{Zhang2015,
+@article{zhang2015,
  title = {Crossing the streams: a framework for streaming analysis of short DNA sequencing reads},
  author = {Zhang, Qingpeng and Awad, Sherine and Brown, C. Titus},
  year = 2015,
@@ -42,7 +42,7 @@ journal = {Proceedings of the National Academy of Sciences}
  doi = {10.7287/peerj.preprints.890v1}
 }
 
-@article{Zhang2014,
+@article{zhang2014,
     author = {Zhang, Qingpeng AND Pell, Jason AND Canino-Koning, Rosangela AND Howe, Adina Chuang AND Brown, C. Titus},
     journal = {PLOS ONE},
     publisher = {Public Library of Science},
