### 2017-06-26 #272 @arfon
> The text already cited those references as is. Perhaps the message appeared because I had not properly included the .bib file and/or included a header for the References section?

The issue was still there I'm afraid - case sensitive references not being picked up. I've fixed up the issues in https://github.com/dib-lab/khmer/pull/1739

When you've merged https://github.com/dib-lab/khmer/pull/1739 could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
