### 2017-06-27 755a7ab
Switch definition of count and add in QFCounttable
This way we can have one implementation and stick to the API
