### 2017-07-01 0088198
Removed import csv from setup.py per betatim's snippet
