### 2017-07-03 #272 @standage
Sorry about the case sensitivity issues, and thanks for your PR. The archive is now posted to Zenodo at [doi:10.5281/zenodo.822478](http://dx.doi.org/10.5281/zenodo.822478).
