### 2017-07-03 #272 @arfon
@sjackman many thanks for your review here and to @biorelated for editing this submission ✨

@standage - your paper is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00272 :zap: 🚀 💥
